split backend_gtkcairo.py into Gcf dependent and independent parts.
